Split LaTeX tests into two files.
