+++ doorstop/core/tests/test_publisher_latex.py
@@ -6,28 +6,18 @@
 
 import os
 import unittest
-from secrets import token_hex
-from shutil import rmtree
 from unittest import mock
 from unittest.mock import Mock, call, patch
 
 from doorstop.common import DoorstopError
 from doorstop.core import publisher
-from doorstop.core.builder import build
 from doorstop.core.document import Document
-from doorstop.core.tests import (
-    ROOT,
-    MockDataMixIn,
-    MockDocument,
-    MockItem,
-    MockItemAndVCS,
-)
+from doorstop.core.tests import MockDataMixIn, MockDocument, MockItem, MockItemAndVCS
 from doorstop.core.tests.helpers_latex import (
     LINES,
     YAML_LATEX_DOC,
     YAML_LATEX_NO_DOC,
     getLines,
-    getWalk,
 )
 from doorstop.core.types import iter_documents
 
@@ -941,81 +931,3 @@ class TestPublisherModule(MockDataMixIn, unittest.TestCase):
         result = getLines(publisher.publish_lines(item, ".tex"))
         # Assert
         self.assertEqual(expected, result)
-
-
-class TestPublisherFullDocument(MockDataMixIn, unittest.TestCase):
-    """Unit tests for the doorstop.core.publisher_latex module by publishing a full document tree."""
-
-    # pylint: disable=no-value-for-parameter
-    def setUp(self):
-        """Setup test folder."""
-        # Build a tree.
-        self.mock_tree = build(cwd=ROOT, root=ROOT, request_next_number=None)
-        self.hex = token_hex()
-        self.dirpath = os.path.join("mock", "LaTeX", self.hex)
-        os.makedirs(self.dirpath)
-        self.expected_walk = """{n}/
-    HLT.tex
-    LLT.tex
-    REQ.tex
-    TUT.tex
-    compile.sh
-    doc-HLT.tex
-    doc-LLT.tex
-    doc-REQ.tex
-    doc-TUT.tex
-    traceability.tex
-    assets/
-        doorstop.cls
-        logo-black-white.png
-""".format(
-            n=self.hex
-        )
-
-    @classmethod
-    def tearDownClass(cls):
-        """Remove test folder."""
-        rmtree("mock")
-
-    def test_publish_latex_tree_copies_assets(self):
-        """Verify that LaTeX assets are published when publishing a tree."""
-        # Act
-        path2 = publisher.publish(self.mock_tree, self.dirpath, ".tex")
-        # Assert
-        self.assertIs(self.dirpath, path2)
-        # Get the exported tree.
-        walk = getWalk(self.dirpath)
-        self.assertEqual(self.expected_walk, walk)
-
-    def test_publish_latex_document_copies_assets(self):
-        """Verify that LaTeX assets are published when publishing a document."""
-        expected_walk = """{n}/
-    TUT.tex
-    compile.sh
-    doc-TUT.tex
-    assets/
-        doorstop.cls
-        logo-black-white.png
-""".format(
-            n=self.hex
-        )
-        # Act
-        dirpath = self.dirpath + "/dummy.tex"
-        path2 = publisher.publish(self.mock_tree.find_document("TUT"), dirpath, ".tex")
-        # Assert
-        self.assertIs(dirpath, path2)
-        # Get the exported tree.
-        walk = getWalk(self.dirpath)
-        self.assertEqual(expected_walk, walk)
-
-    @patch("doorstop.settings.PUBLISH_HEADING_LEVELS", False)
-    def test_publish_document_no_headings_with_latex_data(self):
-        """Verify a LaTeX document can be published with LaTeX doc data but without publishing heading levels."""
-        # Act
-        path2 = publisher.publish(self.mock_tree, self.dirpath, ".tex")
-        # path2 = publisher.publish(self.document, path, ".tex")
-        # Assert
-        self.assertIs(self.dirpath, path2)
-        # Get the exported tree.
-        walk = getWalk(self.dirpath)
-        self.assertEqual(self.expected_walk, walk)
